% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageChildren.R
\name{ageChildren}
\alias{ageChildren}
\title{ageChildren}
\usage{
ageChildren(age, u5mr = 0,
  groups = "6:17=1; 18:29=2; 30:41=3; 42:53=4; 54:59=5")
}
\arguments{
\item{age}{Vector of ages}

\item{u4mr}{Under five years mortality rate as deaths / 10,000 persons / day}
}
\value{
A list of class "ageChildren" with:
    \describe{
      \item{\code{u5mr}}{Under five years mortality rate as deaths / 10000 persons / day}
      \item{\code{observed}}{Table of counts in each (year-centred) age group}
      \item{\code{expected}}{Table of expected counts in each (year-centred) age group}
      \item{\code{X2}}{Chi-squared test statistic}
      \item{\code{df}}{Degrees of freedom for Chi-squared test}
      \item{\code{p}}{p-value for Chi-squared test}
    }
}
\description{
Goodness of fit to an expected (model-based) age distribution with print()
and plot() helper functions
}
\examples{



}
