% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageRatioTest.R
\name{ageRatioTest}
\alias{ageRatioTest}
\title{Age ratio test}
\usage{
ageRatioTest(x, ratio = 0.85)
}
\arguments{
\item{x}{Numeric vector (age)}

\item{ratio}{Expected age ratio}
}
\value{
A lit of class \code{"ageRatioTest"} with:
\describe{
\item{\code{expectedR}}{Expected sex ratio}
\item{\code{expectedP}}{Expected proportion aged 6:29 months}
\item{\code{observedR}}{Observed sex ratio}
\item{\code{observedP}}{Observed proportion aged 6:29 months}
\item{\code{X2}}{Chi-squared test statistic}
\item{\code{df}}{Degrees of freedom for Chi-squared test}
\item{\code{p}}{\code{p-value} for Chi-squared test}
}
}
\description{
Age Ratio Test is an age-related test of survey and data quality.
}
\examples{
# Age-ratio test on survey dataset from Kabul, Afghanistan (dp.ex02)
# with an age ratio of 0.85
svy <- dp.ex02
ageRatioTest(svy$age, ratio = 0.85)

# The age ratio test applied to data for each sex separately
by(svy$age, svy$sex, ageRatioTest, ratio = 0.85)
}
