% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipnTK.R
\docType{data}
\name{as.ex02}
\alias{as.ex02}
\title{Example dataset for age and sex distributions function}
\format{
A data frame with 20 observations and 4 variables
\describe{
\item{\code{age}}{Age group}
\item{\code{Males}}{Total male population}
\item{\code{Females}}{Total female population}
\item{\code{All}}{Total population}
}
}
\source{
\url{http://www.wolframalpha.com/input/?i=Tanzania+age+distribution}
}
\usage{
as.ex02
}
\description{
Census data of Tanzania taken from the Wolfram|Alpha knoweldge engine.
}
\keyword{datasets}
