% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipnTK.R
\docType{data}
\name{dp.ex01}
\alias{dp.ex01}
\title{Example dataset for digit preference function}
\format{A data frame with 796 observations and 6 variables
\describe{
\item{\code{psu}}{Primary sampling unit}
\item{\code{age}}{Age (months)}
\item{\code{sex}}{Gender}
\item{\code{wt}}{Weight (kg)}
\item{\code{ht}}{height (cm)}
\item{\code{oedema}}{Presence or absence of oedema}
}}
\usage{
dp.ex01
}
\description{
Anthropometric data from a single state from a Demographic and Health Survey
(DHS) of a West African country.
}
\keyword{datasets}
