% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipnTK.R
\docType{data}
\name{dp.ex02}
\alias{dp.ex02}
\title{Example dataset for digit preference function}
\format{A data frame with 873 observations and 7 variables
\describe{
\item{\code{psu}}{Primary sampling unit}
\item{\code{age}}{Age of child (months)}
\item{\code{sex}}{Gender of child}
\item{\code{weight}}{Weight of child (kgs)}
\item{\code{height}}{Height of child (cm)}
\item{\code{muac}}{Mid-upper arm circumference (mm)}
\item{\code{oedema}}{Presence or absence of oedema}
}}
\usage{
dp.ex02
}
\description{
Anthropometric data from a SMART survey in Kabul, Afghanistan in a
comma-separated-value (CSV) file format. This is a survey of children aged
6-59 months old.
}
\keyword{datasets}
