% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipnTK.R
\docType{data}
\name{dp.ex03}
\alias{dp.ex03}
\title{Example dataset for digit preference}
\format{
A data frame with 374 observations and 6 variables
\describe{
\item{\code{age}}{Age (months)}
\item{\code{sex}}{Gender (1 = Male / 2 = Female)}
\item{\code{weight}}{Weight (kg)}
\item{\code{height}}{Height (cm)}
\item{\code{muac}}{Mid-upper arm circumference (mm)}
\item{\code{oedema}}{Presence or absence of oedema}
}
}
\usage{
dp.ex03
}
\description{
Anthropometric data for a sample of children living in a refugee camp in a
West African country.
}
\keyword{datasets}
