% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipnTK.R
\docType{data}
\name{flag.ex01}
\alias{flag.ex01}
\title{Example dataset for identifying outliers using flags}
\format{
A data frame with 786 observations and 11 variables
\describe{
\item{\code{psu}}{Primary sampling unit}
\item{\code{child}}{Child ID}
\item{\code{age}}{Age (months)}
\item{\code{sex}}{Gender (1 = Male / 2 = Female)}
\item{\code{weight}}{Weight (kg)}
\item{\code{height}}{Height (cm)}
\item{\code{muac}}{Mid-upper arm circumference (mm)}
\item{\code{oedema}}{Presence or absence of oedema}
\item{\code{haz}}{Height-for-age z-score}
\item{\code{waz}}{Weight-for-age z-score}
\item{\code{whz}}{Weight-for-height z-score}
}
}
\usage{
flag.ex01
}
\description{
Anthropometric data from a SMART survey in Sudan.
}
\keyword{datasets}
