% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipnTK.R
\docType{data}
\name{flag.ex02}
\alias{flag.ex02}
\title{Example dataset for identifying outliers using flags}
\format{A data.frame with 973 observations and 7 variables.
\describe{
\item{\code{school}}{School ID}
\item{\code{sex}}{Gender (1 = Male / 2 = Female)}
\item{\code{ageMonths}}{Age (months)}
\item{\code{weight}}{Weight (kg)}
\item{\code{height}}{Height (cm)}
\item{\code{haz}}{Height-for-age z-score}
\item{\code{baz}}{Body mass index (BMI)-for-age z-score}
}}
\usage{
flag.ex02
}
\description{
Anthropometric data from a survey of children 11 years or older attending
school in Ethiopia.
}
\keyword{datasets}
