% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipnTK.R
\docType{data}
\name{flag.ex03}
\alias{flag.ex03}
\title{Example dataset for identifying outliers using flags}
\format{
A data frame with 18330 observations and 10 variables
\describe{
\item{\code{psu}}{Primary sampling unit}
\item{\code{region}}{Region code}
\item{\code{state}}{State}
\item{\code{age}}{Age (months)}
\item{\code{sex}}{Gender (1 = Male / 2 = Female)}
\item{\code{weight}}{Weight (kg)}
\item{\code{height}}{Height (cm)}
\item{\code{haz}}{Height-for-age z-score}
\item{\code{waz}}{Weight-for-age z-score}
\item{\code{whz}}{Weight-for-height z-score}
}
}
\usage{
flag.ex03
}
\description{
Anthropometric data from a national survey in Nigeria.
}
\keyword{datasets}
