% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipnTK.R
\docType{package}
\name{nipnTK}
\alias{nipnTK}
\alias{nipnTK-package}
\title{NiPN data quality toolkit}
\description{
This is a library of R functions for assessing data-quality in nutritional
anthropometry surveys.
}
