% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliersUV.R
\name{outliersUV}
\alias{outliersUV}
\title{IQR to detect univariate outliers}
\usage{
outliersUV(x, fence = 1.5)
}
\arguments{
\item{x}{Numeric vector}

\item{fence}{\code{IQR} multiplier (defaults to 1.5)}
}
\value{
A logical vector (TRUE for an outlier)
}
\description{
IQR to detect univariate outliers
}
\examples{
# Use outliersUV() to detect univariate outliers in an anthropometric
# dataset from a SMART survey from Angola (rl.ex01)
svy <- rl.ex01
svy[outliersUV(svy$muac), ]

}
