% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageRatioTest.R
\name{print.ageRatioTest}
\alias{print.ageRatioTest}
\title{\code{\link[=print]{print()}} helper function for \code{\link[=ageRatioTest]{ageRatioTest()}} function}
\usage{
\method{print}{ageRatioTest}(x, ...)
}
\arguments{
\item{x}{Object resulting from applying \code{\link[=ageRatioTest]{ageRatioTest()}} function}

\item{...}{Additional \code{\link[=print]{print()}} arguments}
}
\value{
Printed output of \code{\link[=ageRatioTest]{ageRatioTest()}} function
}
\description{
\code{\link[=print]{print()}} helper function for \code{\link[=ageRatioTest]{ageRatioTest()}} function
}
\examples{
# Print age-ratio test results for survey dataset from Kabul, Afghanistan (dp.ex02)
svy <- dp.ex02
print(ageRatioTest(svy$age, ratio = 0.85))

}
