% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digitPreference.R
\name{print.digitPreference}
\alias{print.digitPreference}
\title{\code{print()} helper function for \code{digitPreference()} function}
\usage{
\method{print}{digitPreference}(x, ...)
}
\arguments{
\item{x}{Object resulting from applying the \code{digitPreference()} function.}

\item{...}{Additional \code{print()} parameters}
}
\value{
Printed output of \code{digitPreference()} function
}
\description{
\code{print()} helper function for \code{digitPreference()} function
}
\examples{
# Print output of digit preference test applied to anthropometric data from a
#single state from a DHS survey in a West African country
svy <- dp.ex01
print(digitPreference(svy$wt, digits = 1))
}
