% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexRatioTest.R
\name{print.sexRatioTest}
\alias{print.sexRatioTest}
\title{\code{print()} helper function for \code{sexRatioTest()} function}
\usage{
\method{print}{sexRatioTest}(x, ...)
}
\arguments{
\item{x}{Output resulting from applying the \code{sexRatioTest()} function}

\item{...}{Additional \code{print()} parameters}
}
\value{
Printed output of \code{sexRatioTest()} function
}
\description{
\code{print()} helper function for \code{sexRatioTest()} function
}
\examples{
# Use \\code{sexRatioTest()} on household roster data from a survey in Tanzania
# (as.ex01) and census data of Tanzania extracted from Wolfram|Alpha knowledge
# engine (as.ex02)
svy <- as.ex01
ref <- as.ex02
censusM <- sum(ref$Males)
censusF <- sum(ref$Females)
srt <- sexRatioTest(svy$sex, codes = c(1, 2), pop = c(censusM, censusF))
print(srt)
}
