% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewKurt.R
\name{print.skewKurt}
\alias{print.skewKurt}
\title{\code{print()} helper function for \code{skewKurt()} function}
\usage{
\method{print}{skewKurt}(x, ...)
}
\arguments{
\item{x}{Object resulting from applying the \code{skewKurt()} function}

\item{...}{Additional \code{print()} parameters}
}
\value{
Printed output of \code{skewKurt()} function
}
\description{
\code{print()} helper function for \code{skewKurt()} function
}
\examples{
# Use \code{skewKurt()} on an anthropometric data from a SMART survey in
# Kabul, Afghanistan (dist.ex01)
svy <- dist.ex01
sk <- skewKurt(svy$muac)
print(sk)
}
