% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewKurt.R
\name{skewKurt}
\alias{skewKurt}
\title{Skew and kurtosis}
\usage{
skewKurt(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
A list of class "skewKurt" with:
\describe{
\item{\code{s}}{Skewness with direction}
\item{\code{s.se}}{Standard error of skewness}
\item{\code{s.z}}{Test statistic (\code{s.z = s / s.se})}
\item{\code{s.p}}{\code{p-value} (\code{s != 0})}
\item{\code{k}}{Excess kurtosis with direction}
\item{\code{k.se}}{Standard error of excess kurtosis}
\item{\code{k.z}}{Test statistic (\code{k.z = k / k.se})}
\item{\code{k.p}}{\code{p-value} (\code{k != 0})}
}
}
\description{
Skew and kurtosis
}
\examples{
# Use \\code{skewKurt()} on an anthropometric data from a SMART survey in
# Kabul, Afghanistan (dist.ex01)
svy <- dist.ex01
skewKurt(svy$muac)
}
