% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipnTK.R
\docType{data}
\name{sp.ex01}
\alias{sp.ex01}
\title{Example dataset for using scatterplots to identify outliers}
\format{A data frame with 895 observations and 6 variables
\describe{
\item{\code{age}}{Age (months)}
\item{\code{sex}}{Gender (1 = Male / 2 = Female)}
\item{\code{weight}}{Weight (kg)}
\item{\code{height}}{Height (cm)}
\item{\code{muac}}{Mid-upper arm circumference (mm)}
\item{\code{oedema}}{Presence or absence of oedema}
}}
\usage{
sp.ex01
}
\description{
Anthropometric data from a SMART survey in the Democratic Republic of Congo.
}
\keyword{datasets}
