% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipnTK.R
\docType{data}
\name{sp.ex02}
\alias{sp.ex02}
\title{Example dataset for using scatterplots to identify outliers}
\format{
A data frame with 849 observations and 9 variables
\describe{
\item{\code{region}}{Region code}
\item{\code{school}}{School code}
\item{\code{ageMonths}}{Age (months)}
\item{\code{sex}}{Sex (1 = Male / 2 = Female)}
\item{\code{weight}}{Weight (kg)}
\item{\code{height}}{Height (cm)}
\item{\code{haz}}{Height-for-age z-score}
\item{\code{waz}}{Weight-for-age z-score}
\item{\code{baz}}{Body mass index (BMI)-for-age z-score}
}
}
\usage{
sp.ex02
}
\description{
Anthropometric data from a survey of school-age (i.e., between 5 and 15 years)
children from Pakistan
}
\keyword{datasets}
