% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipnTK.R
\docType{data}
\name{ah.ex01}
\alias{ah.ex01}
\title{Example dataset for age heaping function}
\format{An object of class \code{data.frame} with 593 rows and 10 columns.}
\source{
Data courtesy of \link{HelpAge International}{http://www.helpage.org}
}
\usage{
ah.ex01
}
\description{
Anthropometric data from a Rapid Assessment Method for Older People (RAM-OP)
survey in the Dadaab refugee camp in Garissa, Kenya. This is a survey of
people aged sixty years and older.
}
\keyword{datasets}
