% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ah.ex01}
\alias{ah.ex01}
\title{Example dataset for age heaping function}
\format{
A data frame with 593 observations and 10 variables\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{psu} \tab Primary sampling unit \cr
   \emph{camp} \tab Camp name code \cr
   \emph{block} \tab Block code \cr
   \emph{age} \tab Age (years) \cr
   \emph{sex} \tab Sex \cr
   \emph{weight} \tab Weight (kg) \cr
   \emph{height} \tab Height (cm) \cr
   \emph{demispan} \tab Demispan (cm) \cr
   \emph{muac} \tab Mid-upper arm circumference (cm) \cr
   \emph{oedema} \tab Oedema \cr
}
}
\source{
Data courtesy of \href{https://www.helpage.org}{HelpAge International}
}
\usage{
ah.ex01
}
\description{
Anthropometric data from a Rapid Assessment Method for Older People (RAM-OP)
survey in the Dadaab refugee camp in Garissa, Kenya. This is a survey of
people aged sixty years and older.
}
\keyword{datasets}
