% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{as.ex01}
\alias{as.ex01}
\title{Example dataset for age and sex distributions function}
\format{
A data frame of 8736 observations and 2 variables\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{age} \tab Age (years) \cr
   \emph{sex} \tab Sex (1 = Male / 2 = Female) \cr
}
}
\usage{
as.ex01
}
\description{
Data taken from household rosters collected as part of a household survey in
Tanzania.
}
\keyword{datasets}
