% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{as.ex02}
\alias{as.ex02}
\title{Example dataset for age and sex distributions function}
\format{
A data frame with 20 observations and 4 variables\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{age} \tab Age group \cr
   \emph{Males} \tab Total male population \cr
   \emph{Females} \tab Total female population \cr
   \emph{All} \tab Total population \cr
}
}
\source{
Wolfram|Alpha
}
\usage{
as.ex02
}
\description{
Census data of Tanzania taken from the Wolfram|Alpha knowledge engine.
}
\keyword{datasets}
