% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dp.ex01}
\alias{dp.ex01}
\title{Example dataset for digit preference function}
\format{
A data frame with 796 observations and 6 variables\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{psu} \tab Primary sampling unit \cr
   \emph{age} \tab Age (months) \cr
   \emph{sex} \tab Gender \cr
   \emph{wt} \tab Weight (kg) \cr
   \emph{ht} \tab height (cm) \cr
   \emph{oedema} \tab Presence or absence of oedema \cr
}
}
\usage{
dp.ex01
}
\description{
Anthropometric data from a single state from a Demographic and Health Survey
(DHS) of a West African country.
}
\keyword{datasets}
