% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dp.ex02}
\alias{dp.ex02}
\title{Example dataset for digit preference function}
\format{
A data frame with 873 observations and 7 variables\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{psu} \tab Primary sampling unit \cr
   \emph{age} \tab Age of child (months) \cr
   \emph{sex} \tab Gender of child \cr
   \emph{weight} \tab Weight of child (kgs) \cr
   \emph{height} \tab Height of child (cm) \cr
   \emph{muac} \tab Mid-upper arm circumference (mm) \cr
   \emph{oedema} \tab Presence or absence of oedema \cr
}
}
\usage{
dp.ex02
}
\description{
Anthropometric data from a SMART survey in Kabul, Afghanistan in a
comma-separated-value (CSV) file format. This is a survey of children aged
6-59 months old.
}
\keyword{datasets}
