% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dp.ex03}
\alias{dp.ex03}
\title{Example dataset for digit preference}
\format{
A data frame with 374 observations and 6 variables\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{age} \tab Age (months) \cr
   \emph{sex} \tab Gender (1 = Male / 2 = Female) \cr
   \emph{weight} \tab Weight (kg) \cr
   \emph{height} \tab Height (cm) \cr
   \emph{muac} \tab Mid-upper arm circumference (mm) \cr
   \emph{oedema} \tab Presence or absence of oedema \cr
}
}
\usage{
dp.ex03
}
\description{
Anthropometric data for a sample of children living in a refugee camp in a
West African country.
}
\keyword{datasets}
