% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flag.ex01}
\alias{flag.ex01}
\title{Example dataset for identifying outliers using flags}
\format{
A data frame with 786 observations and 11 variables\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{psu} \tab Primary sampling unit \cr
   \emph{child} \tab Child ID \cr
   \emph{age} \tab Age (months) \cr
   \emph{sex} \tab Gender (1 = Male / 2 = Female) \cr
   \emph{weight} \tab Weight (kg) \cr
   \emph{height} \tab Height (cm) \cr
   \emph{muac} \tab Mid-upper arm circumference (mm) \cr
   \emph{oedema} \tab Presence or absence of oedema \cr
   \emph{haz} \tab Height-for-age z-score \cr
   \emph{waz} \tab Weight-for-age z-score \cr
   \emph{whz} \tab Weight-for-height z-score \cr
}
}
\usage{
flag.ex01
}
\description{
Anthropometric data from a SMART survey in Sudan.
}
\keyword{datasets}
