% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flag.ex02}
\alias{flag.ex02}
\title{Example dataset for identifying outliers using flags}
\format{
A data.frame with 973 observations and 7 variables.\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{school} \tab School ID \cr
   \emph{sex} \tab Gender (1 = Male / 2 = Female) \cr
   \emph{ageMonths} \tab Age (months) \cr
   \emph{weight} \tab Weight (kg) \cr
   \emph{height} \tab Height (cm) \cr
   \emph{haz} \tab Height-for-age z-score \cr
   \emph{baz} \tab Body mass index (BMI)-for-age z-score \cr
}
}
\usage{
flag.ex02
}
\description{
Anthropometric data from a survey of children 11 years or older attending
school in Ethiopia.
}
\keyword{datasets}
