% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flag.ex03}
\alias{flag.ex03}
\title{Example dataset for identifying outliers using flags}
\format{
A data frame with 18330 observations and 10 variables\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{psu} \tab Primary sampling unit \cr
   \emph{region} \tab Region code \cr
   \emph{state} \tab State \cr
   \emph{age} \tab Age (months) \cr
   \emph{sex} \tab Gender (1 = Male / 2 = Female) \cr
   \emph{weight} \tab Weight (kg) \cr
   \emph{height} \tab Height (cm) \cr
   \emph{haz} \tab Height-for-age z-score \cr
   \emph{waz} \tab Weight-for-age z-score \cr
   \emph{whz} \tab Weight-for-height z-score \cr
}
}
\usage{
flag.ex03
}
\description{
Anthropometric data from a national survey in Nigeria.
}
\keyword{datasets}
