% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewKurt.R
\name{print.skewKurt}
\alias{print.skewKurt}
\title{\code{\link[=print]{print()}} helper function for \code{\link[=skewKurt]{skewKurt()}} function}
\usage{
\method{print}{skewKurt}(x, ...)
}
\arguments{
\item{x}{Object resulting from applying the \code{\link[=skewKurt]{skewKurt()}} function}

\item{...}{Additional \code{\link[=print]{print()}} parameters}
}
\value{
Printed output of \code{\link[=skewKurt]{skewKurt()}} function
}
\description{
\code{\link[=print]{print()}} helper function for \code{\link[=skewKurt]{skewKurt()}} function
}
\examples{
# Use skewKurt() on an anthropometric data from a SMART survey in
# Kabul, Afghanistan (dist.ex01)
svy <- dist.ex01
sk <- skewKurt(svy$muac)
print(sk)

}
