% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqNormalPlot.R
\name{qqNormalPlot}
\alias{qqNormalPlot}
\title{Normal quantile-quantile plot}
\usage{
qqNormalPlot(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Normal quantile-quantile plot
}
\examples{
# qqNormalPlot() with data from a SMART survey in Kabul, Afghanistan
# (dist.ex01)
svy <- dist.ex01
qqNormalPlot(svy$muac)
qqNormalPlot(svy$haz)
qqNormalPlot(svy$waz)
qqNormalPlot(svy$whz)

}
