% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rl.ex01}
\alias{rl.ex01}
\title{Example dataset for checking ranges and legal values}
\format{
A data frame with 906 observations and 6 variables\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{age} \tab Age (months) \cr
   \emph{sex} \tab Gender (1 = Male / 2 = Female) \cr
   \emph{weight} \tab Weight (kg) \cr
   \emph{height} \tab Height (cm) \cr
   \emph{muac} \tab Mid-upper arm circumference (mm) \cr
   \emph{oedema} \tab Presence or absence of oedema \cr
}
}
\usage{
rl.ex01
}
\description{
Anthropometric data from a SMART survey in Angola.
}
\keyword{datasets}
