% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp.ex02}
\alias{sp.ex02}
\title{Example dataset for using scatterplots to identify outliers}
\format{
A data frame with 849 observations and 9 variables\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{region} \tab Region code \cr
   \emph{school} \tab School code \cr
   \emph{ageMonths} \tab Age (months) \cr
   \emph{sex} \tab Sex (1 = Male / 2 = Female) \cr
   \emph{weight} \tab Weight (kg) \cr
   \emph{height} \tab Height (cm) \cr
   \emph{haz} \tab Height-for-age z-score \cr
   \emph{waz} \tab Weight-for-age z-score \cr
   \emph{baz} \tab Body mass index (BMI)-for-age z-score \cr
}
}
\usage{
sp.ex02
}
\description{
Anthropometric data from a survey of school-age (i.e., between 5 and 15 years)
children from Pakistan
}
\keyword{datasets}
